+++ apps/fluent-tester/src/E2E/Icon/specs/Icon.spec.macos.ts
@@ -4,20 +4,20 @@ import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('Icon Testing Initialization', function () {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy();
   });
 
-  it('Click and navigate to Icon test page', () => {
+  it('Click and navigate to Icon test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    IconPageObject.scrollToComponentButton();
-    IconPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
+    await IconPageObject.scrollToComponentButton();
+    await IconPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
-    NavigateAppPage.clickAndGoToIconPage();
-    IconPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await NavigateAppPage.clickAndGoToIconPage();
+    await IconPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(IconPageObject.isPageLoaded()).toBeTruthy();
+    await expect(await IconPageObject.isPageLoaded()).toBeTruthy();
   });
 });