+++ apps/fluent-tester/src/E2E/RadioGroup/specs/RadioGroup.spec.win.ts
@@ -11,102 +11,104 @@ import {
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('RadioGroup/RadioButton Testing Initialization', function () {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
   });
 
-  it('Click and navigate to RadioGroup test page', () => {
+  it('Click and navigate to RadioGroup test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    RadioGroupPageObject.scrollToComponentButton();
-    RadioGroupPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
+    await RadioGroupPageObject.scrollToComponentButton();
+    await RadioGroupPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
-    NavigateAppPage.clickAndGoToRadioGroupPage();
-    RadioGroupPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await NavigateAppPage.clickAndGoToRadioGroupPage();
+    await RadioGroupPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(RadioGroupPageObject.isPageLoaded()).toBeTruthy(RadioGroupPageObject.ERRORMESSAGE_PAGELOAD);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
+    await expect(await RadioGroupPageObject.isPageLoaded()).toBeTruthy(RadioGroupPageObject.ERRORMESSAGE_PAGELOAD);
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
   });
 });
 
 describe('RadioGroup/RadioButton Accessibility Testing', () => {
   /* Scrolls and waits for the RadioGroup to be visible on the Test Page */
-  beforeEach(() => {
-    RadioGroupPageObject.scrollToTestElement();
-    RadioGroupPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+  beforeEach(async () => {
+    await RadioGroupPageObject.scrollToTestElement();
+    await RadioGroupPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
   });
 
-  it("Validate RadioGroup's accessibilityRole is correct", () => {
-    expect(RadioGroupPageObject.getAccessibilityRole()).toEqual(RADIOGROUP_A11Y_ROLE);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+  it("Validate RadioGroup's accessibilityRole is correct", async () => {
+    await expect(await RadioGroupPageObject.getAccessibilityRole()).toEqual(RADIOGROUP_A11Y_ROLE);
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it("Validate RadioButton's accessibilityRole is correct", () => {
-    expect(RadioGroupPageObject.getRadioButtonAccesibilityRole()).toEqual(RADIOBUTTON_A11Y_ROLE);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+  it("Validate RadioButton's accessibilityRole is correct", async () => {
+    await expect(await RadioGroupPageObject.getRadioButtonAccesibilityRole()).toEqual(RADIOBUTTON_A11Y_ROLE);
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it('RadioGroup - Set accessibilityLabel', () => {
-    expect(RadioGroupPageObject.getAccessibilityLabel(ComponentSelector.Primary)).toEqual(RADIOGROUP_ACCESSIBILITY_LABEL);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+  it('RadioGroup - Set accessibilityLabel', async () => {
+    await expect(await RadioGroupPageObject.getAccessibilityLabel(ComponentSelector.Primary)).toEqual(RADIOGROUP_ACCESSIBILITY_LABEL);
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it('RadioGroup - Do not set accessibilityLabel -> Default to RadioGroup label', () => {
-    expect(RadioGroupPageObject.getAccessibilityLabel(ComponentSelector.Secondary)).toEqual(RADIOGROUP_TEST_COMPONENT_LABEL);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+  it('RadioGroup - Do not set accessibilityLabel -> Default to RadioGroup label', async () => {
+    await expect(await RadioGroupPageObject.getAccessibilityLabel(ComponentSelector.Secondary)).toEqual(RADIOGROUP_TEST_COMPONENT_LABEL);
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it('RadioButton - Set accessibilityLabel', () => {
-    expect(RadioGroupPageObject.getRBAccessibilityLabel(RadioButtonSelector.First)).toEqual(FIRST_RADIO_BUTTON_ACCESSIBILITY_LABEL);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+  it('RadioButton - Set accessibilityLabel', async () => {
+    await expect(await RadioGroupPageObject.getRBAccessibilityLabel(RadioButtonSelector.First)).toEqual(
+      FIRST_RADIO_BUTTON_ACCESSIBILITY_LABEL,
+    );
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it('RadioButton - Do not set accessibilityLabel -> Default to RadioButton label', () => {
-    expect(RadioGroupPageObject.getRBAccessibilityLabel(RadioButtonSelector.Second)).toEqual(SECOND_RADIO_BUTTON_LABEL);
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+  it('RadioButton - Do not set accessibilityLabel -> Default to RadioButton label', async () => {
+    await expect(await RadioGroupPageObject.getRBAccessibilityLabel(RadioButtonSelector.Second)).toEqual(SECOND_RADIO_BUTTON_LABEL);
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 });
 
-describe('RadioGroup Functional Testing', () => {
+describe('RadioGroup Functional Testing', async () => {
   /* This resets the RadioGroup state by clicking/selecting the 1st RadioButton in the RadioGroup */
-  beforeEach(() => {
-    RadioGroupPageObject.scrollToTestElement();
-    RadioGroupPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+  beforeEach(async () => {
+    await RadioGroupPageObject.scrollToTestElement();
+    await RadioGroupPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
 
-    RadioGroupPageObject.resetRadioGroupSelection();
+    await RadioGroupPageObject.resetRadioGroupSelection();
   });
 
-  it('Click on a RadioButton and ensure it changes state from unselected -> selected', () => {
+  it('Click on a RadioButton and ensure it changes state from unselected -> selected', async () => {
     /* Validate the RadioButton is not initially selected */
-    expect(RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Second)).toBeFalsy();
+    await expect(await RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Second)).toBeFalsy();
 
     /* Click on the RadioButton to select it */
-    RadioGroupPageObject.clickRadioButton(RadioButtonSelector.Second);
-    RadioGroupPageObject.waitForRadioButtonSelected(RadioButtonSelector.Second, PAGE_TIMEOUT);
+    await RadioGroupPageObject.clickRadioButton(RadioButtonSelector.Second);
+    await RadioGroupPageObject.waitForRadioButtonSelected(RadioButtonSelector.Second, PAGE_TIMEOUT);
 
     /* Validate the RadioButton is selected */
-    expect(RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Second)).toBeTruthy();
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+    await expect(await RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Second)).toBeTruthy();
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it('Keyboard to RadioButton and check for Selection state', () => {
+  it('Keyboard to RadioButton and check for Selection state', async () => {
     // Presses the ArrowDown key while the first (A) RadioButton is selected
-    RadioGroupPageObject.sendKey(Keys.Down_Arrow, RadioButtonSelector.First);
-    RadioGroupPageObject.waitForRadioButtonSelected(RadioButtonSelector.Second, 5000);
+    await RadioGroupPageObject.sendKey(Keys.Down_Arrow, RadioButtonSelector.First);
+    await RadioGroupPageObject.waitForRadioButtonSelected(RadioButtonSelector.Second, 5000);
 
     /* Validate the RadioButton is selected */
-    expect(RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Second)).toBeTruthy();
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+    await expect(await RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Second)).toBeTruthy();
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 
-  it("Keyboard to DISABLED RadioButton and validate it doesn't get selected", () => {
+  it("Keyboard to DISABLED RadioButton and validate it doesn't get selected", async () => {
     // Presses the ArrowDown key while the second (B) RadioButton is selected
-    RadioGroupPageObject.sendKey(Keys.Down_Arrow, RadioButtonSelector.Second);
-    RadioGroupPageObject.waitForRadioButtonSelected(RadioButtonSelector.Fourth, 5000); // It should skip RadioButton 3 since it is disabled
+    await RadioGroupPageObject.sendKey(Keys.Down_Arrow, RadioButtonSelector.Second);
+    await RadioGroupPageObject.waitForRadioButtonSelected(RadioButtonSelector.Fourth, 5000); // It should skip RadioButton 3 since it is disabled
 
     /* Validate the RadioButton is selected */
-    expect(RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Fourth)).toBeTruthy();
-    expect(RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
+    await expect(await RadioGroupPageObject.isRadioButtonSelected(RadioButtonSelector.Fourth)).toBeTruthy();
+    await expect(await RadioGroupPageObject.didAssertPopup()).toBeFalsy(RadioGroupPageObject.ERRORMESSAGE_ASSERT);
   });
 });